move input data tab to individual module
